requirements.txt
requests~=2.32.4
qbittorrent-api~=2025.5.0
colorama~=0.4.6
websocket-client~=1.8.0